refactor to get tests to run with trading page mocked queries
