+++ apps/trading-e2e/src/fixtures/markets/markets.json
