+++ apps/trading-e2e/src/integration/deal-ticket.feature
@@ -1,9 +1,6 @@
-Feature: Market orders
-
+Feature: Deal ticket
   Scenario Outline: Successfull market buy orders
-    Given I am on the homepage
-    And I navigate to markets page
-    When I click on active market
+    Given I am on the trading page for an active market 
     And I connect to Vega Wallet
     And place a buy '<marketOrderType>' market order
     Then order request is sent
@@ -14,9 +11,7 @@ Feature: Market orders
       | IOC             |
 
   Scenario Outline: Successfull Limit buy orders
-    Given I am on the homepage
-    And I navigate to markets page
-    When I click on active market
+    Given I am on the trading page for an active market 
     And I connect to Vega Wallet
     And place a buy '<limitOrderType>' limit order
     Then order request is sent
@@ -30,9 +25,7 @@ Feature: Market orders
       | GFN            |
 
   Scenario Outline: Successfull market sell order
-    Given I am on the homepage
-    And I navigate to markets page
-    When I click on active market
+    Given I am on the trading page for an active market 
     And I connect to Vega Wallet
     And place a sell '<marketOrderType>' market order
     Then order request is sent
@@ -43,9 +36,7 @@ Feature: Market orders
       | IOC             |
 
   Scenario Outline: Successfull limit sell order
-    Given I am on the homepage
-    And I navigate to markets page
-    When I click on active market
+    Given I am on the trading page for an active market 
     And I connect to Vega Wallet
     And place a sell '<limitOrderType>' limit order
 
@@ -76,17 +67,13 @@ Feature: Market orders
     And "Market is currently suspended" error is shown
 
   Scenario: Unable to order because wallet is not connected
-    Given I am on the homepage
-    And I navigate to markets page
-    When I click on active market
+    Given I am on the trading page for an active market 
     Then place order button is disabled
     And "No public key selected" error is shown
 
   @ignore
   Scenario: Unsuccessfull because quantity is 0
-    Given I am on the homepage
-    And I navigate to markets page
-    When I click on active market
+    Given I am on the trading page for an active market 
     And I connect to Vega Wallet
     And place a buy 'FOK' market order with amount of 0
     Then Order rejected by wallet error shown containing text "must be positive"
