+++ apps/trading-e2e/src/support/pages/trading-page.js
@@ -1,0 +1,3 @@
+import BasePage from './base-page';
+
+export class TradingPage extends BasePage {}
