+++ apps/trading-e2e/src/support/step_definitions/home-page.step.js
@@ -1,22 +1,23 @@
 import { Then, When } from 'cypress-cucumber-preprocessor/steps';
-import MarketsPage from '../pages/markets-page';
-const marketsPage = new MarketsPage();
+import VegaWallet from '../vega-wallet';
+
+const vegaWallet = new VegaWallet();
 
 When('I try to connect Vega wallet with incorrect details', () => {
-  marketsPage.navigateToConnectVegaWallet();
-  marketsPage.fillInWalletForm('name', 'wrong passphrase');
-  marketsPage.clickConnectVegaWallet();
+  vegaWallet.navigateToConnectVegaWallet();
+  vegaWallet.fillInWalletForm('name', 'wrong passphrase');
+  vegaWallet.clickConnectVegaWallet();
 });
 
 When('I try to connect Vega wallet with blank fields', () => {
-  marketsPage.navigateToConnectVegaWallet();
-  marketsPage.clickConnectVegaWallet();
+  vegaWallet.navigateToConnectVegaWallet();
+  vegaWallet.clickConnectVegaWallet();
 });
 
 Then('wallet not running error message is displayed', () => {
-  marketsPage.validateWalletNotRunningError();
+  vegaWallet.validateWalletNotRunningError();
 });
 
 Then('wallet field validation errors are shown', () => {
-  marketsPage.validateWalletErrorFieldsDisplayed();
+  vegaWallet.validateWalletErrorFieldsDisplayed();
 });
