+++ apps/trading-e2e/src/support/trading-windows/deal-ticket.js
@@ -1,5 +1,4 @@
-import BasePage from './base-page';
-export default class DealTicketPage extends BasePage {
+export default class DealTicket {
   marketOrderType = 'order-type-TYPE_MARKET';
   limitOrderType = 'order-type-TYPE_LIMIT';
   buyOrder = 'order-side-SIDE_BUY';
@@ -53,7 +52,7 @@ export default class DealTicketPage extends BasePage {
     );
     cy.getByTestId(this.orderTransactionHash)
       .invoke('text')
-      .should('contain', 'Tx hash: tx-hash');
+      .should('contain', 'Tx hash: test-tx-hash');
   }
 
   verifyOrderFailedInsufficientFunds() {
@@ -64,9 +63,10 @@ export default class DealTicketPage extends BasePage {
   }
 
   clickPlaceOrder() {
-    if (Cypress.env('bypassPlacingOrders' != true)) {
-      cy.getByTestId(this.placeOrderBtn).click();
-    }
+    // if (Cypress.env('bypassPlacingOrders' != true)) {
+    cy.getByTestId(this.placeOrderBtn).click();
+    cy.contains('Awaiting network confirmation');
+    // }
   }
 
   verifyPlaceOrderBtnDisabled() {
@@ -74,7 +74,10 @@ export default class DealTicketPage extends BasePage {
   }
 
   verifySubmitBtnErrorText(expectedText) {
-    cy.getByTestId(this.inputError).should('have.text', expectedText);
+    cy.getByTestId('dealticket-error-message').should(
+      'have.text',
+      expectedText
+    );
   }
 
   verifyOrderRejected(errorMsg) {
