+++ apps/trading/pages/markets/[marketId].page.tsx
@@ -46,6 +46,7 @@ const MarketPage = () => {
       }}
     >
       {({ market }) => {
+        console.log(market);
         if (!market) {
           return <Splash>{t('Market not found')}</Splash>;
         }
